test(datetime): migrate basic
