+++ core/src/components/datetime/test/basic/datetime.e2e.ts-snapshots/date-highlight-start-of-month-md-ltr-Mobile-Safari-linux.png
